# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NFS API Server",
        "title": "NFS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "nfs-api.andasy.dev",
    "basePath": "/",
    "paths": {
...
        "/api/v1/nla-files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "List all NLA files",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by UPI",
                        "name": "upi",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by seller/buyer name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by office ID",
                        "name": "office_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/utils.PaginatedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "Create a new NLA file",
                "parameters": [
                    {
                        "description": "Create payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nlafile.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/nlafile.NlaFile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nla-files/counts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "Get NLA file status counts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/nlafile.StatusCount"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nla-files/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "Get NLA file by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NLA File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/nlafile.NlaFile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "Update an NLA file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NLA File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/nlafile.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/nlafile.NlaFile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "Delete an NLA file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NLA File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nla-files/{id}/approve": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "Approve an NLA file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NLA File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/nlafile.NlaFile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nla-files/{id}/next": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "Move NLA file to next status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NLA File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/nlafile.NlaFile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nla-files/{id}/reject": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NLA Files"
                ],
                "summary": "Reject an NLA file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NLA File ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/nlafile.NlaFile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "first_name": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "office": {
                    "$ref": "#/definitions/office.Office"
                },
                "office_id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "description": "active | inactive",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "nlafile.CreateRequest": {
            "type": "object",
            "required": [
                "buyer",
                "date",
                "office_id",
                "office_service_id",
                "pending_period_days",
                "seller",
                "upi"
            ],
            "properties": {
                "buyer": {
                    "type": "string",
                    "maxLength": 200
                },
                "comment": {
                    "type": "string"
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "office_id": {
                    "type": "integer"
                },
                "office_service_id": {
                    "type": "integer"
                },
                "pending_period_days": {
                    "type": "integer",
                    "minimum": 1
                },
                "seller": {
                    "type": "string",
                    "maxLength": 200
                },
                "upi": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "nlafile.NlaFile": {
            "type": "object",
            "properties": {
                "approved_at": {
                    "type": "string"
                },
                "buyer": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nla_at": {
                    "type": "string"
                },
                "office": {
                    "$ref": "#/definitions/office.Office"
                },
                "office_id": {
                    "type": "integer"
                },
                "office_service": {
                    "$ref": "#/definitions/officeservice.OfficeService"
                },
                "office_service_id": {
                    "type": "integer"
                },
                "pending_at": {
                    "type": "string"
                },
                "pending_period_days": {
                    "type": "integer"
                },
                "rejected_at": {
                    "type": "string"
                },
                "request_action_at": {
                    "type": "string"
                },
                "scanned_at": {
                    "type": "string"
                },
                "scanning_at": {
                    "type": "string"
                },
                "seller": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "upi": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/auth.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "nlafile.StatusCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "nlafile.UpdateRequest": {
            "type": "object",
            "properties": {
                "buyer": {
                    "type": "string",
                    "maxLength": 200
                },
                "comment": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "office_id": {
                    "type": "integer"
                },
                "office_service_id": {
                    "type": "integer"
                },
                "pending_period_days": {
                    "type": "integer",
                    "minimum": 1
                },
                "seller": {
                    "type": "string",
                    "maxLength": 200
                },
                "upi": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
...
        "office.Office": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "payment_type": {
                    "description": "subscription | full_paid",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "pricing": {
                    "type": "number"
                },
                "scanner_id": {
                    "type": "string"
                },
                "status": {
                    "description": "active | inactive",
                    "type": "string"
                },
                "subscription_end_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "officeservice.OfficeService": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted": {
                    "description": "0 = active, 1 = deleted",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number_of_copies": {
                    "type": "integer"
                },
                "office": {
                    "$ref": "#/definitions/office.Office"
                },
                "office_id": {
                    "type": "integer"
                },
                "pricing": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.PaginatedResponse": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}